requirements.txt
faiss-cpu<2
sentence-transformers
pymupdf
fastapi
uvicorn
pydantic
